requirements.txt
absl-py==1.0.0
alabaster==0.7.12
alembic==1.7.7
astunparse==1.6.3
attrs==21.4.0
av==9.2.0
Babel==2.10.1
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
databricks-cli==0.16.6
docker==5.0.3
docutils==0.17.1
doxypypy==0.8.8.6
easygui==0.98.3
entrypoints==0.4
Flask==2.1.2
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imageio==2.19.0
imagesize==1.3.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.2
mediapipe==0.8.9.1
mlflow==1.25.1
mysql-connector-python==8.0.29
numpy==1.21.6
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pywin32==227
PyYAML==6.0
querystring-parser==1.2.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.36
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tk==0.1.0
tkinter-math==0.2.4
typing_extensions==4.2.0
urllib3==1.26.9
waitress==2.1.1
websocket-client==1.3.2
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
